Update idk
So this is the problem I have, I get an output like:

# GCF_000142965.out
# GCF_000143765.out
# GCF_000143785.out
# GCF_000144995.out
# GCF_000145015.out
Error: [blastdbcmd] Entry not found: WP_005731153.1
Error: [blastdbcmd] Entry or entries not found in BLAST database
Error: [blastdbcmd] Entry not found: WP_005731153.1
Error: [blastdbcmd] Entry or entries not found in BLAST database
# GCF_000145745.out
# GCF_000145845.out
# GCF_000145865.out
Error: [blastdbcmd] Entry not found: WP_005762703.1
Error: [blastdbcmd] Entry or entries not found in BLAST database
Error: [blastdbcmd] Entry not found: WP_005762703.1
Error: [blastdbcmd] Entry or entries not found in BLAST database
Error: [blastdbcmd] Entry not found: WP_004407886.1
Error: [blastdbcmd] Entry or entries not found in BLAST database
Error: [blastdbcmd] Entry not found: WP_004407886.1
Error: [blastdbcmd] Entry or entries not found in BLAST database

So that makes me think that all the protein identifiers I want aren't in BLAST, so we can't use BLAST. BUT, then I go and try to check for the GCF proteins in the 70-CUTGA database (I look up the GCF_xxxxxxx with the missing WP) to see what it looks like and they don't exist in the 70-CUTGA database. So is it not reading the right database?  Also distances aren't printing anymore, I got rid of the prints we had before.